Split EndermuxClient for easier maintenance, add dumb terminal fallback support
